Add new decorators for client/server close.
Currently, it is not possible to set tags on the span based on the
result of a gRPC call (for example, to include the gRPC status code).
Add a new decorator interface to both client and server builders which
allows setting custom tags when the gRPC call is closed.

Fixes #20.