Don't set error tag in interceptor.
Consumers of the gRPC instrumentation may want to have more control over
use of the error tag. For example, a server span which returns NOT_FOUND
may not necessarily be an error server side (it could be expected).
Removing the setting of this tag will give more control to consumers to
determine which gRPC calls are errors or not.