Activate span throughout client call interceptor chain
The ClientInterceptor.interceptCall() method chain returns wrapped ClientCall implementations. Each ClientCall method chain may be executed in a different thread by the client stub. Activating the span in each ClientCall method of the ClientTracingInterceptor ensures the span is available in each ClientCall method chain for all ClientInterceptors that follow the ClientTracingInterceptor. Although the ClientCall.Listener method chain is executed in another thread to the ClientCall method chain, the ClientCall.Listener methods for ClientInterceptors that follow the ClientTracingInterceptor may still access the span by retrieving it in the ClientCall.start() method.