Activate span throughout server call interceptor chain
The ServerInterceptor.interceptCall() method chain returns wrapped ServerCall.Listener implementations. Each ServerCall.Listener method chain may be executed in a different thread by the server. Activating the span in each ServerCall.Listener method of the ServerTracingInterceptor ensures the span is available in each ServerCall.Listener method chain for all ServerInterceptors that follow the ServerTracingInterceptor. The span is also available in the ServerCall methods for ServerInterceptors that follow the ServerTracingInterceptor, since the ServerCall method chain is executed in the same thread as the ServerCall.Listener method chain on the path to and from the server handler.